domashnie_kotiki_bot/requirements.txt
aiogram>=3.13,<4.0
SQLAlchemy>=2.0,<3.0
aiosqlite>=0.20
apscheduler>=3.10,<4.0
python-dotenv>=1.0
aiohttp>=3.9

Pillow>=10.0
